+++ 0x08-recursion/main.h
@@ -9,9 +9,9 @@ void _print_rev_recursion(char *s);
 int _strlen_recursion(char *s);
 int factorial(int n);
 int _pow_recursion(int x, int y);
-int _sqrt_recursion(int n);
+long int _sqrt_recursion(long int n);
 int is_prime_number(int n);
-int sqrt_helper(int n, int start, int end);
+long int sqrt_helper(long int n, long int start, long int end);
 int checkprime(int n, int m);
